used unsigned insterd of size_t

+++ 0x12-singly_linked_lists/1-list_len.c
@@ -10,7 +10,7 @@ Return
 
 size_t list_len(const list_t *h)
 {
-	unsigned int nodes = 0;
+	size_t nodes = 0;
 
 	while (h != NULL)
 	{
